Final commit * minor refactoring, layout fixes

+++ app/src/main/java/com/example/project/database/RestaurantDao.kt
@@ -12,7 +12,7 @@ interface RestaurantDao {
     fun getAll(): LiveData<List<RestaurantData>>
 
     @Query("SELECT * FROM restaurants WHERE id IN (:resID)")
-    suspend fun findByID(resID: Int): RestaurantData?
+    suspend fun findByID(resID: Long): RestaurantData?
 
 
     @Delete
